+++ test/ui-testing/exercise.js
@@ -96,25 +96,16 @@ module.exports.test = (uiTestCtx) => {
               return false;
             }
           })
-          .wait(222)
+          .wait('#input-item-barcode')
           .type('#input-item-barcode', barcode)
-          .wait(222)
+          .wait('#clickable-add-item')
           .click('#clickable-add-item')
-          .wait(1111)
-          .evaluate(() => {
-            const sel = document.querySelector('div[class^="textfieldError"]');
-            if (sel) {
-              throw new Error(sel.textContent);
-            }
-          })
-          .then(() => {
-            nightmare
-              .wait(222)
-              .click('#section-item button')
-              .wait(parseInt(process.env.FOLIO_UI_DEBUG, 10) ? parseInt(config.debug_sleep, 10) : 555) // debugging
-              .then(done);
-          })
-          .catch(done);
+          .wait(`#list-items-checked-out div[title*="${barcode}"]`)
+          .then(done)
+          .catch(e => {
+            console.error(e);
+            done();
+          });
       });
       it(`should find ${barcode} in ${userid}'s open loans`, (done) => {
         nightmare
@@ -124,6 +115,8 @@ module.exports.test = (uiTestCtx) => {
           .wait('#clickable-reset-all')
           .click('#clickable-reset-all')
           .insert('#input-user-search', userid)
+          .wait('button[type=submit]')
+          .click('button[type=submit]')
           .wait(`#list-users div[title="${userid}"]`)
           .click(`#list-users div[title="${userid}"]`)
           .wait('#clickable-viewcurrentloans')
@@ -145,9 +138,9 @@ module.exports.test = (uiTestCtx) => {
       it(`should check in ${barcode}`, (done) => {
         nightmare
           .click('#clickable-checkin-module')
-          .wait(222)
+          .wait('#input-item-barcode')
           .insert('#input-item-barcode', barcode)
-          .wait(222)
+          .wait('#clickable-add-item')
           .click('#clickable-add-item')
           .wait('#list-items-checked-in')
           .evaluate(() => {
@@ -167,6 +160,8 @@ module.exports.test = (uiTestCtx) => {
           .wait('#clickable-reset-all')
           .click('#clickable-reset-all')
           .insert('#input-user-search', userid)
+          .wait('button[type=submit]')
+          .click('button[type=submit]')
           .wait(`div[title="${userid}"]`)
           .click(`div[title="${userid}"]`)
           .wait('#clickable-viewclosedloans')