+++ test/ui-testing/new_proxy.js
@@ -9,6 +9,7 @@ module.exports.test = function foo(uiTestCtx) {
 
     describe('Login > Find user two users > Add proxy to user 1 > Delete sponsor in user 2 > Logout\n', () => {
       let userIds = [];
+      let proxyId = '';
       before((done) => {
         login(nightmare, config, done); // logs in with the default admin credentials
       });
@@ -27,11 +28,12 @@ module.exports.test = function foo(uiTestCtx) {
         nightmare
           .wait('#clickable-users-module')
           .click('#clickable-users-module')
-          .wait(1000)
           .wait('#input-user-search')
-          .wait(1000)
           .type('#input-user-search', '0')
-          .wait(1000)
+          .wait('#clickable-reset-all')
+          .click('#clickable-reset-all')
+          .wait('#clickable-filter-pg-faculty')
+          .click('#clickable-filter-pg-faculty')
           .wait('#list-users div[role="listitem"]:nth-child(1)')
           .evaluate(() => {
             const ubc = [];
@@ -57,34 +59,45 @@ module.exports.test = function foo(uiTestCtx) {
       });
 
       it('should add a proxy for user 1', (done) => {
+        const selector = '#OverlayContainer #list-users div[role="listitem"]:nth-child(1) div[role=gridcell]:nth-child(4)';
         nightmare
-          .wait(1000)
           .wait('#input-user-search')
           .type('#input-user-search', '0')
           .wait('#clickable-reset-all')
           .click('#clickable-reset-all')
           .insert('#input-user-search', userIds[0].barcode)
+          .wait('button[type=submit]')
+          .click('button[type=submit]')
           .wait('#clickable-edituser')
           .click('#clickable-edituser')
           .wait('#accordion-toggle-button-proxy')
           .click('#accordion-toggle-button-proxy')
-          .wait('#proxy button[title^="Find"]')
-          .click('#proxy button[title^="Find"]')
-          .wait('div[aria-label="Select User"] #input-user-search')
-          .insert('div[aria-label="Select User"] #input-user-search', userIds[1].barcode)
-          .wait(222)
-          .wait(`div[aria-label="Select User"] #list-users div[role="listitem"] > a > div[title="${userIds[1].barcode}"]`)
-          .click(`div[aria-label="Select User"] #list-users div[role="listitem"] > a > div[title="${userIds[1].barcode}"]`)
-          .wait('#clickable-updateuser')
-          .click('#clickable-updateuser')
-          .wait(1000)
-          .then(() => {
+          .wait('#proxy #clickable-plugin-find-user')
+          .click('#proxy #clickable-plugin-find-user')
+
+          // you'd think we could just execute a search here,
+          // but clicking the "search" button in the modal submits
+          // the underlying user-edit form in addition to this form,
+          // so we lose the ability to capture the proxy we just found.
+          .wait('#OverlayContainer #clickable-filter-pg-undergrad')
+          .click('#OverlayContainer #clickable-filter-pg-undergrad')
+          .wait('#OverlayContainer #list-users div[role="listitem"]:nth-child(1)')
+          .evaluate((s) => {
+            const pid = document.querySelector(s).innerText;
+            return pid;
+          }, selector)
+          .then(text => {
+            nightmare
+              .click('#OverlayContainer #list-users div[role="listitem"]:nth-child(1) a')
+              .wait('#clickable-updateuser')
+              .click('#clickable-updateuser');
             done();
+            proxyId = text;
           })
           .catch(done);
       });
 
-      it('should delete a sponsor of user 2', (done) => {
+      it(`should delete a sponsor of user 2 (${proxyId})`, (done) => {
         nightmare
           .wait(2222)
           // put some junk in the search field to get the reset button
@@ -93,12 +106,11 @@ module.exports.test = function foo(uiTestCtx) {
           .type('#input-user-search', 'asdf')
           .wait('#clickable-reset-all')
           .click('#clickable-reset-all')
-          .wait(222)
-          .type('#input-user-search', userIds[1].barcode)
-          .wait(`#list-users div[role="listitem"] > a > div[title="${userIds[1].barcode}"]`)
-          .wait(222)
-          .click(`#list-users div[role="listitem"] > a > div[title="${userIds[1].barcode}"]`)
-          .wait(222)
+          .type('#input-user-search', proxyId)
+          .wait('button[type=submit]')
+          .click('button[type=submit]')
+          .wait(`#list-users div[role="listitem"] > a > div[title="${proxyId}"]`)
+          .click(`#list-users div[role="listitem"] > a > div[title="${proxyId}"]`)
           .wait('#accordion-toggle-button-proxySection')
           .wait('#clickable-edituser')
           .click('#clickable-edituser')
@@ -106,13 +118,10 @@ module.exports.test = function foo(uiTestCtx) {
           .click('#accordion-toggle-button-proxy')
           .wait(`#proxy a[href*="${userIds[0].uuid}"]`)
           .xclick(`id("proxy")//a[contains(@href, "${userIds[0].uuid}")]/../../../..//button`)
-          .wait(2111)
           .wait('#clickable-deleteproxies-confirmation-confirm')
-          .wait(2111)
           .click('#clickable-deleteproxies-confirmation-confirm')
           .wait('#clickable-updateuser')
           .click('#clickable-updateuser')
-          .wait(1111)
           .then(() => { done(); })
           .catch(done);
       });