+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change history for platform-core
 
-## 1.0.0 (IN PROGRESS)
+## 1.2.0 (IN PROGRESS)
+
+* Manually click the search button in tests. Refs STCOM-354.
+* Don't use the search button in OverlayContainers; that submits the underlying forms in addition.
+
+
+## [1.1.0](https://github.com/folio-org/platform-core/tree/v1.1.0) (2018-10-09)
+* Upgrade platform and all modules to versions targeting stripes framework 1.0, FOLIO-1547
+
+
+## [1.0.9](https://github.com/folio-org/platform-core/tree/v1.0.9) (2018-09-26)
+* 
+
+
+## 1.0.0
 * New platform created for building and testing core modules
 * Move cross-module tests from `ui-testing` into `platform-core`, UITEST-22
 * Update package scripts to pass working directory token with --run, UITEST-36